# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicAI SaaS multi-tenant WhatsApp automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/professionals:
    get:
      operationId: listProfessionals
      tags: [professionals]
      summary: List professionals for a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: serviceId
          in: query
          required: false
          schema:
            type: integer
          description: Filter professionals who can perform this service
      responses:
        "200":
          description: List of professionals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professional"
    post:
      operationId: createProfessional
      tags: [professionals]
      summary: Create a professional
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalBody"
      responses:
        "201":
          description: Created professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
  /clinics/{clinicId}/professionals/{id}:
    get:
      operationId: getProfessional
      tags: [professionals]
      summary: Get a professional with their service links
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professional detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionalDetail"
    patch:
      operationId: updateProfessional
      tags: [professionals]
      summary: Update a professional
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessionalBody"
      responses:
        "200":
          description: Updated professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
    delete:
      operationId: deleteProfessional
      tags: [professionals]
      summary: Delete a professional
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clinics/{clinicId}/professionals/{id}/services:
    put:
      operationId: setProfessionalServices
      tags: [professionals]
      summary: Set the services a professional can perform (replaces existing)
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetProfessionalServicesBody"
      responses:
        "200":
          description: Updated service list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionalDetail"
  /clinics/{clinicId}/professionals/{id}/schedule:
    get:
      operationId: getProfessionalSchedule
      tags: [professionals]
      summary: Get the weekly schedule of a professional
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of schedule entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfessionalScheduleEntry"
    put:
      operationId: setProfessionalSchedule
      tags: [professionals]
      summary: Replace the full weekly schedule of a professional
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetProfessionalScheduleBody"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfessionalScheduleEntry"
...
components:
  schemas:
...
    Professional:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clinicId, name, specialty, active, createdAt, updatedAt]
    ProfessionalDetail:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        serviceIds:
          type: array
          items:
            type: integer
      required: [id, clinicId, name, specialty, active, createdAt, updatedAt, serviceIds]
    CreateProfessionalBody:
      type: object
      properties:
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        active:
          type: boolean
        serviceIds:
          type: array
          items:
            type: integer
      required: [name, specialty]
    UpdateProfessionalBody:
      type: object
      properties:
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        active:
          type: boolean
    SetProfessionalServicesBody:
      type: object
      properties:
        serviceIds:
          type: array
          items:
            type: integer
      required: [serviceIds]
    ProfessionalScheduleEntry:
      type: object
      properties:
        id:
          type: integer
        professionalId:
          type: integer
        dayOfWeek:
          type: integer
          description: 0=Sunday, 1=Monday … 6=Saturday
        startMinute:
          type: integer
          description: Minutes since midnight (e.g. 480 = 08:00)
        endMinute:
          type: integer
          description: Minutes since midnight (e.g. 1020 = 17:00)
        isBlock:
          type: boolean
          description: false = work window, true = blocked slot (e.g. lunch break)
        createdAt:
          type: string
          format: date-time
      required: [id, professionalId, dayOfWeek, startMinute, endMinute, isBlock, createdAt]
    SetProfessionalScheduleBody:
      type: object
      properties:
        entries:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: integer
              startMinute:
                type: integer
              endMinute:
                type: integer
              isBlock:
                type: boolean
            required: [dayOfWeek, startMinute, endMinute, isBlock]
      required: [entries]
...